# Simplify locale setup barrier