--- a/modules/internal/ChapelLocale.chpl
+++ b/modules/internal/ChapelLocale.chpl
@@ -626,16 +626,21 @@ module ChapelLocale {
     // initialization (e.g., rootLocale).
     iter chpl_initOnLocales(param tag: iterKind)
       where tag==iterKind.standalone {
-      // Simple locales barrier, see implementation below for notes
-      var b: localesBarrier;
-      var flags: [1..#numLocales-1] unmanaged localesSignal?;
-      coforall locIdx in 0..#numLocales /*ref(b)*/ {
+      // Split into 2 coforalls to barrier after the yield. Ideally, we would
+      // just use a real barrier, but `chpl_comm_barrier` is in use and other
+      // custom barriers have been non-scalable in the past.
+      coforall locIdx in 0..#numLocales {
         on __primitive("chpl_on_locale_num",
                        chpl_buildLocaleID(locIdx:chpl_nodeID_t,
                                           c_sublocid_any)) {
           chpl_defaultDistInitPrivate();
           yield locIdx;
-          b.wait(locIdx, flags);
+        }
+      } // Relying on barrier at join, do NOT fuse these loops
+      coforall locIdx in 0..#numLocales  {
+        on __primitive("chpl_on_locale_num",
+                       chpl_buildLocaleID(locIdx:chpl_nodeID_t,
+                                          c_sublocid_any)) {
           chpl_rootLocaleInitPrivate(locIdx);
           chpl_defaultLocaleInitPrivate();
           chpl_singletonCurrentLocaleInitPrivate(locIdx);
@@ -657,70 +662,6 @@ module ChapelLocale {
       c_free(p);
     }
   }
-
-  //
-  // Simple locales barrier
-  //
-  // This sits outside of the abstract root locale definition above
-  // because the compiler cannot resolve the constructor (known issues
-  // with nested classes/records).  In addition, we cannot have the
-  // flags array in the record, because the initCopy function needs
-  // the dataPar* configs declared in ChapelDistribution.
-  //
-  // Each non-0 locale increments the count, and then waits on a
-  // *local* atomic.  This is done by creating an array of type
-  // class localesSignal, one per locale, and allocating each locale's
-  // copy before updating the count.  Locale 0 waits for the others to
-  // arrive and then set the signals to true.  We can't do anything
-  // too complicated this early on, so we are using a for loop to
-  // broadcast that we are done.
-  pragma "no doc"
-  class localesSignal {
-    var s: atomic bool;
-
-    // Override default initializer; this could go away when the
-    // compiler's default initializer creates a version that takes a
-    // bool rather than an 'atomic bool' (which generates a
-    // '--warn-unstable' warning otherwise)
-    proc init() {
-    }
-  }
-  pragma "no doc"
-  record localesBarrier {
-    proc wait(locIdx, flags) {
-      if locIdx==0 {
-        // locale 0 has nothing else to do, so check flags
-        while (true) {
-          // This fence ensures that writes to the count variables
-          // are available to this task. (Note that they aren't
-          // atomic if they're 128-bit writes though - so we
-          // have some risk of getting part of a wide pointer).
-          // Without this fence, there is a race condition on
-          // a weakly-ordered architecture.
-          atomicFence();
-          var count = 0;
-          for f in flags do
-            if f then count += 1;
-          if count==numLocales-1 then break;
-          // Give time to other tasks/threads/processes
-          // like we do in waitFor
-          chpl_task_yield();
-        }
-        // Let the others go
-        for f in flags do
-          f!.s.testAndSet();
-      } else {
-        var f = new unmanaged localesSignal();
-        // expose my flag to locale 0
-        flags[locIdx] = f;
-        // wait (locally) for locale 0 to set my flag
-        f.s.waitFor(true);
-        // clean up
-        delete f;
-      }
-    }
-  }
-
 
   // This function is called in the LocaleArray module to initialize
   // the rootLocale.  It sets up the origRootLocale and also includes