--- a/test/modules/sungeun/init/printInitCommCounts.good
+++ b/test/modules/sungeun/init/printInitCommCounts.good
@@ -1,1 +1,1 @@
-(execute_on = 4, execute_on_fast = 4, execute_on_nb = 4) (get = 13, put = 3, execute_on_fast = 6) (get = 13, put = 3, execute_on_fast = 6) (get = 13, put = 3, execute_on_fast = 6) (get = 13, put = 3, execute_on_fast = 6)
+(execute_on_fast = 4, execute_on_nb = 8) (get = 10, put = 1, execute_on_fast = 7) (get = 10, put = 1, execute_on_fast = 7) (get = 10, put = 1, execute_on_fast = 7) (get = 10, put = 1, execute_on_fast = 7)