--- a/test/modules/sungeun/init/printInitCommCounts.lm-numa.good
+++ b/test/modules/sungeun/init/printInitCommCounts.lm-numa.good
@@ -1,1 +1,0 @@
-(execute_on = 4, execute_on_fast = 4, execute_on_nb = 4) (get = 125, put = 3, execute_on_fast = 6) (get = 125, put = 3, execute_on_fast = 6) (get = 125, put = 3, execute_on_fast = 6) (get = 125, put = 3, execute_on_fast = 6)