--- a/test/visibility/except/operatorsExceptions/exceptGreaterThanOrEqual.good
+++ b/test/visibility/except/operatorsExceptions/exceptGreaterThanOrEqual.good
@@ -6,4 +6,4 @@ $CHPL_HOME/modules/internal/ChapelBase.chpl:nnnn: note: is passed to formal 'a: 
 exceptGreaterThanOrEqual.chpl:8: note: other candidates are:
 $CHPL_HOME/modules/internal/ChapelBase.chpl:nnnn: note:   >=(a: enum, b: enum)
 $CHPL_HOME/modules/internal/ChapelBase.chpl:nnnn: note:   >=(a: uint(64), b: int(64))
-note: and 64 other candidates, use --print-all-candidates to see them
+note: and 63 other candidates, use --print-all-candidates to see them