@ronawho 2021-12-15
Noticed this recently while working on https://github.com/Cray/chapel-private/issues/2484#issuecomment-993857430. I had been wanting to do something like this for a while but couldn't get `chpl_comm_barrier()` freed up so I gave up on it.

This solution seems really simple and easy in hindsight so I'm worried that I'm missing something and doing something obviously wrong hence the high number of reviewers requested. I believe David reviewed the original PR so pinging him and I'm wondering if Brad might have as well. Pinging Engin for general barrier and locale model set up experience.