@bradcray 2021-12-15
This seems clever and attractive from the amount of special-to-bootstrap startup code it seems
to remove (and the obvious perf impact).  The rewrite seems sound to me assuming that
`on __primitive("chpl_on_locale_num",` has similar synchronization/fence semantics as a normal
`on-clause` (though maybe the fence part doesn't matter if the coforall takes care of that anyway?).
I expect it would, but am not so familiar with it, so that's the one part that made me pause.