@ronawho 2021-12-15
> The rewrite seems sound to me assuming that on __primitive("chpl_on_locale_num", has similar synchronization/fence semantics as a normal on-clause 

It's a "normal" on-clause, we just can't do `on Locales[locIdx]` because this is the code that's setting up the locales array. That primitive is just building up a targetable locale from an idx but the MCM/sync semantics are normal.